create build + release dockerfiles, cleanup examples
