+++ .dockerignore
@@ -1,3 +1,1 @@
-*
-!amd64/surreal
-!arm64/surreal
+target
