+++ lib/Cargo.toml
@@ -15,6 +15,9 @@ categories = ["database-implementations", "data-structures", "embedded"]
 license = "Apache-2.0"
 resolver = "2"
 
+[lib]
+path = "src/lib.rs"
+
 [features]
 # Public features
 default = ["protocol-ws", "rustls"]
@@ -115,6 +118,7 @@ features = [
 ]
 
 [dev-dependencies]
+actix-web = { version = "4.3.0", features = ["macros"] }
 time = { version = "0.3.17", features = ["serde"] }
 tokio = { version = "1.25.0", features = ["macros", "rt", "rt-multi-thread"] }
 
@@ -129,3 +133,23 @@ ws_stream_wasm = "0.7.4"
 tokio = { version = "1.25.0", default-features = false, features = ["io-util", "fs", "rt-multi-thread"] }
 tokio-tungstenite = { version = "0.18.0", optional = true }
 uuid = { version = "1.3.0", features = ["serde", "v4", "v7"] }
+
+[[example]]
+name = "concurrency"
+path = "examples/concurrency.rs"
+
+[[example]]
+name = "query"
+path = "examples/query.rs"
+
+[[example]]
+name = "select"
+path = "examples/select.rs"
+
+[[example]]
+name = "transaction"
+path = "examples/transaction.rs"
+
+[[example]]
+name = "version"
+path = "examples/version.rs"
