--- lib/examples/actix/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "actix-example"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-actix-web = { version = "4.3.0", features = ["macros"] }
-serde = { version = "1.0.152", features = ["derive"] }
-surrealdb = { path = "../.." }
-thiserror = "1.0.38"
