+++ lib/examples/query.rs
