+++ lib/examples/select.rs
