+++ lib/examples/version.rs
